# Conflicts:
#	src/arcgis/apps/_url_schemes.py

+++ src/arcgis/apps/workforce/project.py
@@ -1,13 +1,15 @@
 """ Defines the Project object.
 """
 import arcgis
-from arcgis.features import FeatureLayer
+from arcgis.features import FeatureLayer, Table
 from arcgis.gis import Group
 from arcgis._impl.common._utils import _lazy_property
+from warnings import warn
 import json
 
 from ._schemas import *
 from .managers import *
+from arcgis.apps.workforce.exceptions import WorkforceError
 
 
 class Project:
@@ -19,6 +21,10 @@ class Project:
     ------------------     --------------------------------------------------------------------
     item                   Required :class:`~arcgis.gis.Item`. The item that
                            the contains the project.
+                           
+                           For a version 1 Workforce project, this is an item of type
+                           `Workforce Project`. For a version 2 Workforce project, this is an
+                           item of type `Feature Service` with typeKeyword `Workforce Project`
     ==================     ====================================================================
 
     .. code-block:: python
@@ -31,6 +37,10 @@ class Project:
         project = arcgis.apps.workforce.Project(item)
         assignments = project.assignments.search()
         workers = project.workers.search()
+        
+        # Create v1 "Classic" Workforce project and v2 "offline-enabled" project
+        v1_project = arcgis.apps.workforce.create_project('v1_project', major_version=1)
+        v2_project = arcgis.apps.workforce.create_project('v2_project', major_version=2)
 
 
     """
@@ -40,11 +50,23 @@ class Project:
             :param item: The project's arcigs.gis.Item
         """
         self.gis = item._gis
-        self._item = item
-        self._item_data = item.get_data()
+        if "Workforce Project" in item.typeKeywords:
+            self._item = item
+        else:
+            raise WorkforceError("Incorrect item passed into Project class")
+        if self._is_v2_project:
+            self._item_data = item.properties
+        else:
+            self._item_data = item.get_data()
         self._assignment_schema = AssignmentSchema(self.assignments_layer)
-        self._track_schema = TrackSchema(self.tracks_layer)
+        if self._supports_tracks:
+            self._track_schema = TrackSchema(self.tracks_layer)
+        else:
+            self._track_schema = None
         self._worker_schema = WorkerSchema(self.workers_layer)
+        if self._is_v2_project:
+            self._assignment_types = AssignmentTypeSchema(self.assignment_types_table)
+            self._integration_schema = IntegrationSchema(self.integrations_table)
         self._dispatcher_schema = DispatcherSchema(self.dispatchers_layer)
         self._update_cached_objects()
 
@@ -67,40 +89,45 @@ class Project:
                 self._cached_dispatcher = d
                 break
         else:
-            raise Exception("'{}' is not a dispatcher, please authenticate as a dispatcher".format(self.gis.users.me.username))
+            raise WorkforceError("'{}' is not a dispatcher, please authenticate as a dispatcher".format(self.gis.users.me.username))
 
     def __str__(self):
         return self.title
 
     def __repr__(self):
-        return "<Project {}>".format(self.id)
+        return "<Project {}>".format(self.title)
 
     def delete(self):
         """
             Deletes the project, group, folder, layers, and webmaps.
             Assumes the currently signed in user owns the project or is an admin.
         """
+        title = self.title
         owner = self._item.owner
         self.assignments_item.protect(False)
         self.assignments_item.delete()
-        self.tracks_item.protect(False)
-        self.tracks_item.delete()
-        self.workers_item.protect(False)
-        self.workers_item.delete()
-        self.dispatchers_item.protect(False)
-        self.dispatchers_item.delete()
+        if self._supports_tracks:
+            self.tracks_item.protect(False)
+            self.tracks_item.delete()
+        if not self._is_v2_project:
+            self.workers_item.protect(False)
+            self.workers_item.delete()
+            self.dispatchers_item.protect(False)
+            self.dispatchers_item.delete()
+            self._item.protect(False)
+            self._item.delete()
         self.dispatcher_webmap.item.protect(False)
         self.dispatcher_webmap.item.delete()
         self.worker_webmap.item.protect(False)
         self.worker_webmap.item.delete()
-        self._item.protect(False)
-        self._item.delete()
+        self.group.protected = False
         self.group.delete()
         for folder in self.gis.users.get(owner).folders:
-            if folder['id'] == self._item_data['folderId']:
+            if folder['title'] == title:
                 self.gis.content.delete_folder(folder['title'], owner=owner)
 
     def _update_data(self):
+        # this function is used by v1 projects only
         self._item.update({
             "text": json.dumps(self._item_data)
         })
@@ -119,24 +146,42 @@ class Project:
         if summary:
             item_properties['snippet'] = summary
         self._item.update(item_properties)
+        if self._is_v2_project:
+            self.gis.content.get(self.dispatcher_web_map_id).update(item_properties)
+            self.gis.content.get(self.worker_web_map_id).update(item_properties)
+
+    @property
+    def _supports_tracks(self):
+        return not self._is_v2_project and "tracks" in self._item_data
+
+    @property
+    def _is_v2_project(self):
+        try:
+            return self._item.type == "Feature Service" and int(self._item.properties['workforceProjectVersion'].split(".")[0]) >= 2
+        except:
+            return False
 
     @property
     def _tracking_enabled(self):
-        return self._item_data["tracks"]["enabled"]
+        if self._supports_tracks:
+            return self._item_data["tracks"]["enabled"]
 
     @_tracking_enabled.setter
     def _tracking_enabled(self, value):
-        self._item_data["tracks"]["enabled"] = value
-        self._update_data()
+        if self._supports_tracks:
+            self._item_data["tracks"]["enabled"] = value
+            self._update_data()
 
     @property
     def _tracking_interval(self):
-        return self._item_data["tracks"]["updateInterval"]
+        if self._supports_tracks:
+            return self._item_data["tracks"]["updateInterval"]
 
     @_tracking_interval.setter
     def _tracking_interval(self, value):
-        self._item_data["tracks"]["updateInterval"] = value
-        self._update_data()
+        if self._supports_tracks:
+            self._item_data["tracks"]["updateInterval"] = value
+            self._update_data()
 
     @property
     def integrations(self):
@@ -170,57 +215,139 @@ class Project:
     @property
     def version(self):
         """The version of the project"""
-        return self._item_data['version']
+        if self._is_v2_project:
+            return self._item_data['workforceProjectVersion']
+        else:
+            return self._item_data['version']
 
     @_lazy_property
     def assignments_item(self):
         """The assignments :class:`~arcgis.gis.Item`"""
-        return self.gis.content.get(self._item_data['assignments']['serviceItemId'])
+        if self._is_v2_project:
+            # this is the same item as workers_item, dispatchers_item for a v2 project - each points to the one FS
+            return self.gis.content.get(self._item.id)
+        else:
+            return self.gis.content.get(self._item_data['assignments']['serviceItemId'])
 
     @property
     def assignments_layer_url(self):
         """The assignments feature layer url"""
-        return self._item_data['assignments']['url']
+        if self._is_v2_project:
+            return self._item.url + "/0"
+        else:
+            return self._item_data['assignments']['url']
+    
+    @_lazy_property
+    def assignment_types_item(self):
+        """The assignment types :class:`~arcgis.gis.Item`"""
+        if self._is_v2_project:
+            return self.gis.content.get(self._item.id)
+        else:
+            warn("This Workforce Project does not have an assignment types item", WorkforceWarning)
+
+    @property
+    def assignment_types_table_url(self):
+        """The assignment types table url"""
+        if self._is_v2_project:
+            return self._item.url + "/3"
+        else:
+            warn("This Workforce Project does not have an assignment types table", WorkforceWarning)
 
     @_lazy_property
     def dispatchers_item(self):
         """The dispatchers :class:`~arcgis.gis.Item`"""
-        return self.gis.content.get(self._item_data['dispatchers']['serviceItemId'])
+        if self._is_v2_project:
+            return self.gis.content.get(self._item.id)
+        else:
+            return self.gis.content.get(self._item_data['dispatchers']['serviceItemId'])
 
     @property
     def dispatchers_layer_url(self):
-        """The dispatchers feature layer url"""
-        return self._item_data['dispatchers']['url']
+        """The dispatchers layer url"""
+        if self._is_v2_project:
+            return self._item.url + "/2"
+        else:
+            return self._item_data['dispatchers']['url']
 
+    @_lazy_property
+    def integrations_table_url(self):
+        """The integrations table url :class:`~arcgis.features.Table`"""
+        if self._is_v2_project:
+            return self._item.url + "/4"
+        else:
+            warn("This Workforce Project does not have an integrations table", WorkforceWarning)
+            
     @_lazy_property
     def tracks_item(self):
         """The tracks :class:`~arcgis.gis.Item`"""
-        return self.gis.content.get(self._item_data['tracks']['serviceItemId'])
+        if self._supports_tracks:
+            return self.gis.content.get(self._item_data['tracks']['serviceItemId'])
+        else:
+            warn("This Workforce Project does not support tracks.", WorkforceWarning)
 
     @property
     def tracks_layer_url(self):
         """The tracks feature layer url"""
-        return self._item_data['tracks']['url']
+        if self._supports_tracks:
+            return self._item_data['tracks']['url']
+        else:
+            warn("This Workforce Project does not support tracks.", WorkforceWarning)
 
+    @_lazy_property
+    def workers_item(self):
+        """The workers :class:`~arcgis.gis.Item`"""
+        if self._is_v2_project:
+            return self.gis.content.get(self._item.id)
+        else:
+            return self.gis.content.get(self._item_data['workers']['serviceItemId'])
+    
     @property
     def workers_layer_url(self):
         """The workers feature layer url"""
-        return self._item_data['workers']['url']
+        if self._is_v2_project:
+            return self._item.url + "/1"
+        else:
+            return self._item_data['workers']['url']
 
     @property
     def dispatcher_web_map_id(self):
         """The dispatcher webmap item id"""
-        return self._item_data['dispatcherWebMapId']
+        if self._is_v2_project:
+            # not all systems will support this new "Workforce2MapFeatureService" so we try/except. If the rel does
+            # not exist, we can get the webmap out of the metadata
+            try:
+                related_items = self._item.related_items('WorkforceMap2FeatureService', 'reverse')
+                for item in related_items:
+                    if "Workforce Dispatcher" in item.typeKeywords:
+                        return item.id
+                return self._item_data['workforceDispatcherMapId']
+            except Exception:
+                return self._item_data['workforceDispatcherMapId']
+        else:
+            return self._item_data['dispatcherWebMapId']
 
     @property
     def worker_web_map_id(self):
         """The worker webmap item id"""
-        return self._item_data['workerWebMapId']
+        if self._is_v2_project:
+            try:
+                related_items = self._item.related_items('WorkforceMap2FeatureService', 'reverse')
+                for item in related_items:
+                    if "Workforce Worker" in item.typeKeywords:
+                        return item.id
+                return self._item_data['workforceWorkerMapId']
+            except Exception:
+                return self._item_data['workforceWorkerMapId']
+        else:
+            return self._item_data['workerWebMapId']
 
     @property
     def group_id(self):
         """The group id that all project items are part of"""
-        return self._item_data['groupId']
+        if self._is_v2_project:
+            return self._item_data['workforceProjectGroupId']
+        else:
+            return self._item_data['groupId']
 
     @_lazy_property
     def owner(self):
@@ -235,17 +362,34 @@ class Project:
     @_lazy_property
     def dispatchers_layer(self):
         """The dispatchers :class:`~arcgis.features.FeatureLayer`"""
-        return FeatureLayer(self.dispatchers_layer_url, self.gis)
+        if self._is_v2_project:
+            return Table(self.dispatchers_layer_url, self.gis)
+        else:
+            return FeatureLayer(self.dispatchers_layer_url, self.gis)
 
+    @_lazy_property
+    def assignment_types_table(self):
+        """The assignment types :class:`~arcgis.features.Table`"""
+        if self._is_v2_project:
+            return Table(self.assignment_types_table_url, self.gis)
+        else:
+            warn("This Workforce Project does not have an assignment types table", WorkforceWarning)
+            
+    @_lazy_property
+    def integrations_table(self):
+        """The integrations :class:`~arcgis.features.Table`"""
+        if self._is_v2_project:
+            return Table(self.integrations_table_url, self.gis)
+        else:
+            warn("This Workforce Project does not have an integrations table", WorkforceWarning)
+            
     @_lazy_property
     def tracks_layer(self):
         """The tracks :class:`~arcgis.features.FeatureLayer`"""
-        return FeatureLayer(self.tracks_layer_url, self.gis)
-
-    @_lazy_property
-    def workers_item(self):
-        """The workers :class:`~arcgis.gis.Item`"""
-        return self.gis.content.get(self._item_data['workers']['serviceItemId'])
+        if self._supports_tracks:
+            return FeatureLayer(self.tracks_layer_url, self.gis)
+        else:
+            warn("This Workforce Project does not support tracks.", WorkforceWarning)
 
     @_lazy_property
     def workers_layer(self):
@@ -285,7 +429,10 @@ class Project:
     @property
     def tracks(self):
         """The :class:`~arcgis.apps.workforce.managers.TrackManager` for the project"""
-        return TrackManager(self)
+        if self._supports_tracks:
+            return TrackManager(self)
+        else:
+            warn("This Workforce Project does not support tracks.", WorkforceWarning)
 
     @property
     def assignment_types(self):
